Sign the `pkg-config` executable.
Add a code signing target for the `pkg-config` executable.  Since the
executable is signed in-place, create stamp file
`$(TMP)/signed.stamp.txt` to track the signing state.  Added the
makefile variable `APP_SIGNING_ID` to hold the name of the signing
certificate on the Keychain.

Added info about the `APP_SIGNING_ID` variable to the readme file.

Also added additional steps to the `check` target: verifying the
executable's signature and verifying the installer package's signature.