Add `notarize` target to the makefile.
Add new makefile variable `NOTARIZATION_KEYCHAIN_PROFILE` to hold the
Keychain name of the notarization credentials.

Rename the default target from `all` to `signed-package`, since it's not
really "all" anymore.

Add two extra checks to the `check` target: run `spctl --assess --type
install` on the installer to check its notarization status, and run
`xtapler validate` on the installer to check that the notarization
ticket has been stapled to it.

Modified the `configure` flags to add `--disable-host-tool` to remove
the unneeded `$(host)-pkg-config` hard link to the `pkg-config`
executable (where `host` is something like `arm64-apple-darwin20`).
This extra copy of the executable isn't really needed, and would be
incorrectly named when built on arm and installed on x86_64 or
vice-versa.

The notarization targets track this sequence of files:
- `pkg-config-$(version).pkg` --> `$(TMP)/submit-log.json` tracks the
  submission to the notarization service
- `$(TMP)/submit-log.json` --> `$(TMP)/submission-id.txt` extracts the
  submission ID
- `$(TMP)/submission-id.txt` --> `$(TMP)/notarization-log.json`
  retrieves notarization details
- `$(TMP)/notarization-log.json` --> `$(TMP)/notarization.stamp.txt`
  checks if the installer as accepted (i.e. notariztion success)
- `$(TMP)/notarized.stamp.txt`` --> `$(TMP)/stapled.stamp.txt` tracks
  stapling status of the installer

Also updated the readme file, adding a section on prerequisites, adding
the `NOTARIZATION_KEYCHAIN_PROFILE` variable, a section on the `make
notarize` target, and a section on the required signing and notarizing
credentials.